Fix a bug about success_rate_is_high_enough (it wasn't updated correctly, so some samples didn't rerun); add SampleSuccessMonitor to clean up the code

+++ examples/goldilocks_models/goldilocks_utils.cc
@@ -36,7 +36,7 @@ SubQpData::SubQpData(int N_sample) {
     ub_vec.push_back(std::make_shared<VectorXd>());
     y_vec.push_back(std::make_shared<VectorXd>());
     B_vec.push_back(std::make_shared<MatrixXd>());
-    is_success_vec.push_back(std::make_shared<int>());
+    is_success_vec.push_back(std::make_shared<double>());
 
     A_active_vec.push_back(std::make_shared<MatrixXd>());
     B_active_vec.push_back(std::make_shared<MatrixXd>());
